Changed the diff logic to only use the "baseCommit" field for submitted reviews.
This is important, because for a pending review, what the user really cares about is
the current diff between their review branch and their target branch, not the historical
diff between what their target branch *was* and what their review branch currently is.

To provide a more concrete example; let's say you start a review, see that there are
new changes in your target branch, merge in those changes, and then run
"git appraise show --diff".

In that scenario, you do not want your diff to include the latest changes to the
target branch that you have merged in. This change makes it so that such changes
will not be displayed as part of the diff.