Changed the handling of the git repo to be injected rather than an implicit singleton.
This, in turn, will allow us to build a mock git repo that we can use for testing,
as opposed to the current test behavior, which relies on this repos review history.

To keep this change manageable, it only includes the injecting of the repo. The
updates to the tests will come later.