BREAKING CHANGE: Improve consistency of the command line interface.
This change makes the CLI consistent for all of the review subcommands.

Specifically:
1. The usage strings for every command follow a consistent format.
2. The "comment" command now takes the file and line as flags, and allows
   the review hash to be provided as a positional argument.

The change to the "comment" command is backwards-incompatible. Scripts
that relied on being able to pass the file and line number as positional
arguments will no longer work with this new version.

To mitigate the issues with that; the previous commit has been tagged
as version "v0.1", so that users who need the previous behavior can always
install the older version of the tool.