Changed the logic that calls into git to always cascade errors upwards.
Previously, we treated some errors as fatal if they were unrecoverable, and
used log.Fatal(...) to kill the program.

That old behavior was desirable when the code was simply being used by the
command-line tool, as it allowed us to catch unexpected errors very easily.
However, now that this code is being reused as a library, we should let the
caller decide how to treat unrecoverable errors.