Populate the default commit message when submitting a review with the --merge option.
The "git appraise submit" command supports a "--merge" option, that forces the
creation of a merge commit for the review (sometimes called a "merge bubble").

When this is done, an editor is launched for the user to provide the message
for that merge commit. Previously, the editor was being populated with the
default message which has the form "merge branch ... with ...". This change
replaces that with a more appropriate message that specifies the review
being submitted, and the description of that review.

This is important in that a user may want to use merge commits to allow
themselves to both maintain a full history of their changes (including changes
made in response to review feedback) and have the ability to list only the
high level changes corresponding to the submission of each review.

More specifically, if a user wants to always submit their reviews with the
--merge option, then this change makes it so that they can use the command
"git log --first-parent" to view only the portion of their history that
corresponds to review submissions, while still having the displayed messages
provide a meaningful summary of the change.