Improve the error messages returned when a git command fails.
Previously, we were simply cascading the error returned by the instance
of exec.Command. However, that winds up just being something of the form
"exit status 128", with all of the real error message going to the
Stderr field.

As such, this commit changes the behavior to save the data written to
stderr, and use it to construct a new error to return.