Add the ability to submit an arbitrary review.
Previously, we were assuming that the review requester would submit the
review, so the submit command required that you check out the review ref
before you could submit. However, with pull requests a requester might
not have permission to submit a review. As such, we should make it easy
for a reviewer to submit a review.

This change allows that scenario, by adding support for specifying the
hash of the review to submit.