Split out the summary of a review into a separate struct.
This new struct is then returned by all of the methods that list the
reviews in a repo, and is what the "list" subcommand operates on.

By making this split, we can separate the data that is cheap to look up
from the data that is expensive to look up, and forgoe the expensive
look ups when we are operating on multiple reviews.

This change improves the performance of listing the open reviews in the
git-appraise repo by more than three fold (dropping the run time from
more than 7 seconds to less than 2).