Formalize how to handle multiple review requests on a commit.
Previously, the tool was taking the last request in a note as being
the current one, and relying on the combination of appending notes
and merging them using "cat_sort_uniq" to make that correspond to
the request with the latest timestamp.

However, that was neither as robust as it could be, nor was it
documented in our spec.

This change makes the tool pick the last request by timestamp, and
updates the spec to indicate that this is how the notes should be
interpreted.

This change also makes the sorting of both requests and comments stable,
and extends the Summary struct to store all of the review requests in
addition to storing the current one.

This change is a (partial) response to feedback in
https://github.com/google/git-appraise/issues/33.