Ignoring error for git config appraise.submit
If an empty string error message is returned when git config appraise.submit
is ran, git-appraise will ignore it and default to using the fast-forward
strategy on review submission.

The following flags may override that behaviour:

--merge
--rebase
--fast-forward

For example:

$ git config --add appraise.submit 'merge'
$ git appraise submit --rebase j8hgs74skf

would rebase rather than merge. Had git config not been ran it would still
rebase and had rebase not been set as a flag in this case along with no
git config, it would fast-forward.

Instances of mergeStrategy have also been changed to submitStrategy. Error
messages will still appear for other git config related commands as of #37
when an empty string is returned.