git-appraise without subcommand will list reviews
As stated in #42, this will now show open reviews and not the usage text.
I'm still a little unsure about this, because the list command is obviously
still there and will work with flags, -a and --json for example.

Should this (non-subcommand version) allow the same, and if so, do we even need
the list subcommand?

/cc @jishi9