Removed erroneous inclusion of the "--ancestry-path" flag.
We specifically specify that the `from` parameter of the
`ListCommitsBetween` method should not need to be an ancestor of
the `to` parameter.

However, the arguments to the `git rev-list` call were requiring
that the `from` be an ancestor of the `to`, or else no commits
would be returned.

It turns out that the only calls to `ListCommitsBetween` from
git-appraise itself were forcing the `from` parameter to be an
ancestor of the `to` parameter, so this bug was not manifesting
itself in the tool. However, it did show up when git-appraise was
being used as a library by other tools.