Accepting -F flag on git-appraise comments
I've gathered a few thoughts in regards to #56 in the form of this commit. This
currently doesn't take in to account stdin (-), just a file name.

Basically, the user has 3 choices as this point:

  - 1 git appraise comment -m "nice code!" HASH
  - 2 git appraise comment HASH
  - 3 git appraise comment -F message HASH

Given that 'message' is the file name that contains the message. As for the
stdin, keeping that similar to git would be a good idea. They present the user
with the following output and then read:

```
if (isatty(0))
    fprintf(stderr, _("(reading log message from standard input)\n"));
read_from_stdin(&log);
```

This would be consistent and imo seems like it could work well. I agree with
@ojarjur in that this should be supported with accept, comment, reject
and request. This commit only takes into account, comment. Once that's nailed
down, applying it to the rest would be trivial.