Add the concept of an archive of old review commits (for rebased reviews)
This change adds an "archive" ref (named 'refs/archives/devtools') that
is used to keep track of commits that are part of the review history,
but that will not become part of the final repository history (because
they have been replaced during a rebase operation).

That new ref will be pushed and pulled during invocations of `git
appraise push` and `git appraise pull` so that the full review history
continues to be available to anyone who performs a `git appraise pull`
of the repository.

This change also fixes a bug in the submit workflow where the rebase
command was being run with the incorrect argument.