Better support for rebasing reviews
This change adds the ability to track review changes across rebases.

This means that a review can be commented upon, rewritten
completely, and then submitted, and we will still be able to
keep track of all the intermediate commits, comments, and the
final submission status of the review.

This support is based on a new field in the request schema
called `alias`. When populated, that field indicates the hash
of commit that is the result of rebasing the review.