Correctly compute the base commit for abandoned reviews.
This fixes an issue where the base commit for abandoned reviews
could not be computed (and thus, the diff could not be shown)
because abandoned reviews were being treated as open rather than
closed.

This fixes that issue by making abandoned reviews be treated as
closed.

This also adds a new API method IsOpen to the 'review' package
because I noticed that the underlying logic encapsulated by that
method was being repeated several times.