[WIP] Implement full source file range support
The review format from ShipeShape supports fulle specification
of a block of text withint a file. We implement this by allowing
the cli to take a range specified as:

`startLine[+startColumn][:endLine[+endColumn]]`

We need to fully check the semantics of the specified location.
Lines and columns are number from 1 -> max(uint32). The start must
be before the start of the file.