Fix a bug in the choice of comment hash.
This fixes a bug in displaying comment threads where the hash shown
for the thread (which is what a user would use for replying) was
the hash of the current version of the comment, rather than the
hash of the original comment.

The hash of the original comment was already being correctly picked
as the hash of the thread, but the code was incorrectly displaying
something else in its place.