Add coloring for review state in "git appraise list"
This commit adds different color highlighting to different review states
in the output of "git-appraise list".

The default colors are:
- abandon: magenta,
- accepted: green,
- danger: yellow red bold blink,
- pending: cyan,
- rejected: yellow red bold strike,
- submitted: yellow,
- tbr: red white bold blink.

They can be modified by the user by setting the corresponding
configuration options in the color.appraise section. E.g. to use magenta
for the "pending" state, set color.appraise.pending to "magenta":
 [color "appraise"]
     pending = magenta