Make the PR-mirroring workflow use the entire history of the current branch
The default behavior of the GitHub workflow is to only fetch the last
commit of the current branch (resulting in a shallow repository).

This causes issues for the workflow to mirror pull request metadata
into git-notes, as the full history of the repository is necessary
in order for that to work.

This change fixes that by modifying the local respitory to no longer
be a shallow clone.